Moved FallbackMonitor back to utilities
