+++ platform-sdk/consensus-roster/build.gradle.kts
@@ -18,7 +18,6 @@ testModuleInfo {
     requires("org.hiero.base.utility.test.fixtures")
     requires("org.hiero.consensus.platformstate")
     requires("org.hiero.consensus.roster.test.fixtures")
-    requires("org.hiero.consensus.utility.test.fixtures")
     requires("org.assertj.core")
     requires("org.junit.jupiter.api")
     requires("org.junit.jupiter.params")
@@ -26,5 +25,4 @@ testModuleInfo {
     requires("org.mockito.junit.jupiter")
 
     runtimeOnly("com.swirlds.config.impl")
-    runtimeOnly("org.hiero.consensus.utility")
 }
