+++ platform-sdk/consensus-roster/src/main/java/module-info.java
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
 module org.hiero.consensus.roster {
-    exports org.hiero.consensus.monitoring;
     exports org.hiero.consensus.roster.schemas;
     exports org.hiero.consensus.roster;
 
@@ -13,7 +12,5 @@ module org.hiero.consensus.roster {
     requires transitive org.hiero.consensus.model;
     requires com.swirlds.base;
     requires org.hiero.base.utility;
-    requires org.hiero.consensus.metrics;
-    requires org.hiero.consensus.utility;
     requires static transitive com.github.spotbugs.annotations;
 }
